updated projects
I designed this website for Malini at Affinity Therapy Purley. Malini had a strong brand and theme, which I was able to create. Malini told me she would like to include nature, so I used an image as a background, which reflects the calm and gentle nature of Malini and her way of working.